Update to put ssr instance in its own file
