+++ packages/web/src/ssr/track/+onBeforeRender.tsx
@@ -1,33 +1,8 @@
 import { Maybe } from '@audius/common/utils'
-import {
-  sdk,
-  full as FullSdk,
-  DiscoveryNodeSelector,
-  productionConfig,
-  stagingConfig,
-  developmentConfig
-} from '@audius/sdk'
+import { full as FullSdk } from '@audius/sdk'
 import type { PageContextServer } from 'vike/types'
 
-const sdkConfigs = {
-  production: productionConfig,
-  staging: stagingConfig,
-  development: developmentConfig
-}
-
-const discoveryNodeSelector = new DiscoveryNodeSelector({
-  bootstrapServices: (
-    sdkConfigs[process.env.VITE_ENVIRONMENT as keyof typeof sdkConfigs] ??
-    productionConfig
-  ).discoveryNodes
-})
-
-export const audiusSdk = sdk({
-  appName: process.env.VITE_PUBLIC_HOSTNAME ?? 'audius.co',
-  services: {
-    discoveryNodeSelector
-  }
-})
+import { audiusSdk } from '../sdk'
 
 export type TrackPageProps = {
   track: Maybe<FullSdk.TrackFull>
